remove page index in post list view

+++ zSMTH/libSMTH/SMTHHelper.h
@@ -41,9 +41,9 @@
 - (BOOL) removeFavorite:(NSString*)engName;
 
 - (NSArray*) getGuidancePosts:(int)section;
-- (NSArray*) getPostsFromBoard:(NSString*)boardID from:(int)from;
+- (NSArray*) getPostsFromBoard:(NSString*)boardID from:(long)from;
 - (NSArray*) getPostContents:(NSString *)board_id postID:(long)article_id from:(long)from;
-- (NSArray*) getFilteredPostsFromBoard:(NSString*)boardID title:(NSString*)title user:(NSString*)user from:(int)from;
+- (NSArray*) getFilteredPostsFromBoard:(NSString*)boardID title:(NSString*)title user:(NSString*)user from:(long)from;
 
 
 - (SMTHUser*) getUserInfo:(NSString*) userID;
